Refactor: move user domain classes to `carp.common`.
