+++ carp.common/src/commonMain/kotlin/dk/cachet/carp/common/serialization/Serialization.kt
@@ -1,10 +1,26 @@
 package dk.cachet.carp.common.serialization
 
+import dk.cachet.carp.common.users.AccountIdentity
+import dk.cachet.carp.common.users.EmailAccountIdentity
+import dk.cachet.carp.common.users.UsernameAccountIdentity
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonConfiguration
 import kotlinx.serialization.modules.EmptyModule
+import kotlinx.serialization.modules.plus
 import kotlinx.serialization.modules.SerialModule
+import kotlinx.serialization.modules.SerializersModule
 
+
+/**
+ * Types in the [dk.cachet.carp.common] module which need to be registered when using [Json] serializer.
+ */
+val COMMON_SERIAL_MODULE = SerializersModule {
+    polymorphic( AccountIdentity::class )
+    {
+        UsernameAccountIdentity::class with UsernameAccountIdentity.serializer()
+        EmailAccountIdentity::class with EmailAccountIdentity.serializer()
+    }
+}
 
 /**
  * Create a [Json] serializer adopting a default CARP infrastructure configuration.
@@ -20,5 +36,5 @@ fun createDefaultJSON( module: SerialModule = EmptyModule ): Json
         // Therefore, continuing to use array polymorphism (as it was in older 'kotlinx.serialization' versions) is preferred.
         useArrayPolymorphism = true )
 
-    return Json( configuration, module )
+    return Json( configuration, COMMON_SERIAL_MODULE + module )
 }
