+++ carp.common/src/commonMain/kotlin/dk/cachet/carp/common/users/AccountIdentity.kt
@@ -1,4 +1,4 @@
-package dk.cachet.carp.deployment.domain.users
+package dk.cachet.carp.common.users
 
 import dk.cachet.carp.common.EmailAddress
 import kotlinx.serialization.Polymorphic
