+++ carp.deployment.core/src/commonMain/kotlin/dk/cachet/carp/deployment/application/UserService.kt
@@ -2,9 +2,9 @@ package dk.cachet.carp.deployment.application
 
 import dk.cachet.carp.common.EmailAddress
 import dk.cachet.carp.common.UUID
-import dk.cachet.carp.deployment.domain.users.Account
+import dk.cachet.carp.common.users.Account
+import dk.cachet.carp.common.users.Username
 import dk.cachet.carp.deployment.domain.users.Participant
-import dk.cachet.carp.deployment.domain.users.Username
 
 
 /**
