+++ carp.deployment.core/src/commonMain/kotlin/dk/cachet/carp/deployment/infrastructure/Serialization.kt
@@ -6,30 +6,15 @@ import dk.cachet.carp.common.serialization.createDefaultJSON
 import dk.cachet.carp.deployment.domain.MasterDeviceDeployment
 import dk.cachet.carp.deployment.domain.StudyDeploymentSnapshot
 import dk.cachet.carp.deployment.domain.StudyDeploymentStatus
-import dk.cachet.carp.deployment.domain.users.Account
-import dk.cachet.carp.deployment.domain.users.AccountIdentity
-import dk.cachet.carp.deployment.domain.users.EmailAccountIdentity
+import dk.cachet.carp.common.users.Account
+import dk.cachet.carp.common.users.Username
 import dk.cachet.carp.deployment.domain.users.Participant
-import dk.cachet.carp.deployment.domain.users.Username
-import dk.cachet.carp.deployment.domain.users.UsernameAccountIdentity
 import dk.cachet.carp.protocols.infrastructure.PROTOCOLS_SERIAL_MODULE
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.modules.EmptyModule
 import kotlinx.serialization.modules.plus
 import kotlinx.serialization.modules.SerialModule
-import kotlinx.serialization.modules.SerializersModule
 
-
-/**
- * Types in the [dk.cachet.carp.deployment] module which need to be registered when using [Json] serializer.
- */
-val DEPLOYMENT_SERIAL_MODULE = SerializersModule {
-    polymorphic( AccountIdentity::class )
-    {
-        UsernameAccountIdentity::class with UsernameAccountIdentity.serializer()
-        EmailAccountIdentity::class with EmailAccountIdentity.serializer()
-    }
-}
 
 /**
  * Create a [Json] serializer adopting a default CARP infrastructure configuration with all [dk.cachet.carp.deployment] types registered.
@@ -38,7 +23,7 @@ val DEPLOYMENT_SERIAL_MODULE = SerializersModule {
  */
 fun createDeploymentSerializer( module: SerialModule = EmptyModule ): Json
 {
-    return createDefaultJSON( PROTOCOLS_SERIAL_MODULE + DEPLOYMENT_SERIAL_MODULE + module )
+    return createDefaultJSON( PROTOCOLS_SERIAL_MODULE + module )
 }
 
 /**
@@ -52,37 +37,37 @@ var JSON: Json = createDeploymentSerializer()
  * Create a [Account] from JSON, serialized using the globally set infrastructure serializer ([JSON]).
  */
 fun Account.Companion.fromJson( json: String ): Account =
-        JSON.parse( serializer(), json )
+    JSON.parse( serializer(), json )
 
 /**
  * Serialize to JSON, using the globally set infrastructure serializer ([JSON]).
  */
 fun Account.toJson(): String =
-        JSON.stringify( Account.serializer(), this )
+    JSON.stringify( Account.serializer(), this )
 
 /**
  * Create a [Username] from JSON, serialized using the globally set infrastructure serializer ([JSON]).
  */
 fun Username.Companion.fromJson( json: String ): Username =
-        JSON.parse( serializer(), json )
+    JSON.parse( serializer(), json )
 
 /**
  * Serialize to JSON, using the globally set infrastructure serializer ([JSON]).
  */
 fun Username.toJson(): String =
-        JSON.stringify( Username.serializer(), this )
+    JSON.stringify( Username.serializer(), this )
 
 /**
  * Create a [Participant] from JSON, serialized using the globally set infrastructure serializer ([JSON]).
  */
 fun Participant.Companion.fromJson( json: String ): Participant =
-        JSON.parse( serializer(), json )
+    JSON.parse( serializer(), json )
 
 /**
  * Serialize to JSON, using the globally set infrastructure serializer ([JSON]).
  */
 fun Participant.toJson(): String =
-        JSON.stringify( Participant.serializer(), this )
+    JSON.stringify( Participant.serializer(), this )
 
 /**
  * Create a [StudyDeploymentSnapshot] from JSON, serialized using the globally set infrastructure serializer ([JSON]).
