+++ carp.deployment.core/src/commonMain/kotlin/dk/cachet/carp/deployment/infrastructure/UserServiceRequest.kt
@@ -5,10 +5,10 @@ import dk.cachet.carp.common.UUID
 import dk.cachet.carp.common.ddd.createServiceInvoker
 import dk.cachet.carp.common.ddd.createServiceInvokerOverloaded
 import dk.cachet.carp.common.ddd.ServiceInvoker
+import dk.cachet.carp.common.users.Account
+import dk.cachet.carp.common.users.Username
 import dk.cachet.carp.deployment.application.UserService
-import dk.cachet.carp.deployment.domain.users.Account
 import dk.cachet.carp.deployment.domain.users.Participant
-import dk.cachet.carp.deployment.domain.users.Username
 import kotlinx.serialization.Serializable
