+++ carp.deployment.core/src/commonTest/kotlin/dk/cachet/carp/deployment/domain/users/InMemoryUserRepository.kt
@@ -1,6 +1,8 @@
 package dk.cachet.carp.deployment.domain.users
 
 import dk.cachet.carp.common.UUID
+import dk.cachet.carp.common.users.Account
+import dk.cachet.carp.common.users.AccountIdentity
 
 
 /**
@@ -9,6 +11,7 @@ import dk.cachet.carp.common.UUID
 class InMemoryUserRepository : UserRepository
 {
     private val accounts: MutableList<Account> = mutableListOf()
+    private val participations: MutableMap<UUID, MutableSet<Participant>> = mutableMapOf()
 
 
     override fun addAccount( account: Account )
@@ -29,21 +32,15 @@ class InMemoryUserRepository : UserRepository
         val account = accounts.firstOrNull { it.id == accountId }
         require( account != null )
 
-        // Add participant to cloned account.
-        val updatedParticipations = account.studyParticipations.plus( participant )
-        val newAccount = account.copy( studyParticipations = updatedParticipations )
-
-        // Remove and re-add account.
-        accounts.remove( account )
-        accounts.add( newAccount )
+        val accountParticipations = participations.getOrPut( accountId ) { mutableSetOf() }
+        accountParticipations.add( participant )
     }
 
+    override fun getStudyParticipations( accountId: UUID ): List<Participant> =
+        participations[ accountId ]?.toList() ?: listOf()
+
     override fun getParticipantsForStudy( studyId: UUID ): List<Participant> =
-        accounts.fold( mutableListOf() ) { participants, account ->
-            val participations = account.studyParticipations.filter { it.studyId == studyId }
-            participants.addAll( participations )
-            participants
-        }
+        participations.flatMap { it.component2().filter { p -> p.studyId == studyId } }
 }
