+++ carp.deployment.core/src/commonTest/kotlin/dk/cachet/carp/deployment/infrastructure/UserServiceRequestsTest.kt
@@ -3,9 +3,9 @@ package dk.cachet.carp.deployment.infrastructure
 import dk.cachet.carp.common.EmailAddress
 import dk.cachet.carp.common.UUID
 import dk.cachet.carp.common.ddd.ServiceInvoker
+import dk.cachet.carp.common.users.Username
 import dk.cachet.carp.deployment.application.UserService
 import dk.cachet.carp.deployment.application.UserServiceMock
-import dk.cachet.carp.deployment.domain.users.Username
 import dk.cachet.carp.test.runBlockingTest
 import kotlin.test.*
